pyproject.toml
[project]
name = "crew-scheduling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "einops>=0.8.1",
    "ipykernel>=6.30.1",
    "numpy>=2.0.2",
    "pyyaml>=6.0.2",
    "tensordict>=0.10.0",
    "torch>=2.8.0",
]
